chore: upgrade deps + rename folder
