+++ next.config.js
@@ -2,6 +2,14 @@ const WpAutoImport = require("unplugin-auto-import/webpack").default;
 
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  images: {
+    remotePatterns: [
+      {
+        protocol: "https",
+        hostname: "img.clerk.com",
+      },
+    ],
+  },
   webpack(config) {
     config.plugins.push(
       WpAutoImport({
