+++ package.json
@@ -10,7 +10,7 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "@clerk/nextjs": "4.27.2",
+    "@clerk/nextjs": "4.27.4",
     "@radix-ui/react-accordion": "^1.1.2",
     "@radix-ui/react-checkbox": "^1.0.4",
     "@radix-ui/react-collapsible": "^1.0.3",
@@ -30,25 +30,25 @@
     "date-fns": "^2.30.0",
     "lucide-react": "^0.294.0",
     "next": "14.0.3",
-    "react": "^18.2.0",
+    "react": "18.3.0-canary-aba93acb6-20231205",
     "react-day-picker": "^8.9.1",
-    "react-dom": "^18.2.0",
-    "react-use": "17.4.1",
-    "tailwind-merge": "^2.0.0",
+    "react-dom": "18.3.0-canary-aba93acb6-20231205",
+    "react-use": "17.4.2",
+    "tailwind-merge": "^2.1.0",
     "tailwindcss-animate": "^1.0.7",
     "zustand": "4.4.7"
   },
   "devDependencies": {
-    "@biomejs/biome": "1.4.0",
-    "@types/node": "^20.10.0",
-    "@types/react": "^18.2.39",
+    "@biomejs/biome": "1.4.1",
+    "@types/node": "^20.10.3",
+    "@types/react": "^18.2.42",
     "@types/react-dom": "^18.2.17",
     "autoprefixer": "^10.4.16",
     "husky": "8.0.3",
-    "postcss": "^8.4.31",
-    "tailwindcss": "^3.3.5",
-    "typescript": "^5.3.2",
-    "unplugin-auto-import": "0.17.1"
+    "postcss": "^8.4.32",
+    "tailwindcss": "^3.3.6",
+    "typescript": "^5.3.3",
+    "unplugin-auto-import": "0.17.2"
   },
   "packageManager": "yarn@4.0.1"
 }
